pyproject.toml
[project]
name = "tfl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.10.1",
    "python-dotenv>=1.1.1",
]

[project.scripts]
tfl = "tfl:main"
journey = "journey:main"
